refactor: move proposal draft outreach service into outreach package
Aligns the new service with main's research_ai services structure.
